fix: fixed build & vercel errors
