+++ src/app/[locale]/layout.tsx
@@ -1,22 +1,22 @@
-"use client"
-
-import { I18nProviderClient, useCurrentLocale } from "../../../locales/client"
 import "@/app/globals.css"
 import Header from "@/components/header/Header"
 import Footer from "@/components/footer/Footer"
+import LocaleWrapper from "./LocaleWrapper"
+import { Metadata } from "next"
 
-export default function LocaleLayout({
+export const metadata: Metadata = {
+  title: "Darrell Roberts - Frontend Engineer",
+  description:
+    "Portfolio website of Darrell Roberts. I'm a Frontend Engineer who lives in Berlin, Germany with a focus on Frontend. I work primarily with TypeScript on the following frameworks or libraries: React, Next, Vue and Nuxt. However, I also work with Node.js in the Backend and have experience with MongoDB and creating my own APIs.",
+}
+
+export default async function LocaleLayout({
   children,
+  params,
 }: {
   children: React.ReactNode
-  params: { locale: string }
+  params: Promise<{ locale: string }>
 }) {
-  const locale = useCurrentLocale()
-  return (
-    <I18nProviderClient locale={locale}>
-      <Header />
-      {children}
-      <Footer />
-    </I18nProviderClient>
-  )
+  const { locale } = await params
+  return <LocaleWrapper locale={locale}>{children}</LocaleWrapper>
 }
