+++ src/components/contact/ContactHome.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import Link from "next/link"
 import React, { useContext } from "react"
 import ContactLinks from "./ContactLinks"
